+++ be/sidedish/src/main/resources/schema.sql
@@ -33,7 +33,7 @@ CREATE TABLE SIDEDISH_ITEM
 CREATE TABLE SIDEDISH_EVENT
 (
     ID          BIGINT PRIMARY KEY AUTO_INCREMENT,
-    EVENT_NAME VARCHAR(50) NOT NULL,
+    EVENT_NAME  VARCHAR(50) NOT NULL,
     EVENT_COLOR VARCHAR(10) NOT NULL
 );
 
@@ -46,17 +46,17 @@ CREATE TABLE SIDEDISH_EVENT_ITEM
 
 CREATE TABLE SIDEDISH_IMAGE
 (
-    ID             BIGINT PRIMARY KEY AUTO_INCREMENT,
-    IMAGE_URL  VARCHAR(200) NOT NULL
+    ID        BIGINT PRIMARY KEY AUTO_INCREMENT,
+    IMAGE_URL VARCHAR(200) NOT NULL
 
 );
 
 CREATE TABLE SIDEDISH_ITEM_IMAGE
 (
-    ID                    BIGINT PRIMARY KEY AUTO_INCREMENT,
-    SIDEDISH_IMAGE     BIGINT REFERENCES SIDEDISH_IMAGE (ID),
+    ID                BIGINT PRIMARY KEY AUTO_INCREMENT,
+    SIDEDISH_IMAGE    BIGINT REFERENCES SIDEDISH_IMAGE (ID),
     SIDEDISH_ITEM     BIGINT REFERENCES SIDEDISH_ITEM (ID),
-    IMAGE_TYPE VARCHAR(15) NOT NULL,
+    IMAGE_TYPE        VARCHAR(15) NOT NULL,
     SIDEDISH_ITEM_KEY INT
 );
